feat: [TG-37] Code register screen
