+++ app/build.gradle
@@ -9,7 +9,7 @@ plugins {
 
 def keystorePropertiesFile = rootProject.file("keystore.properties")
 def keystoreProperties = new Properties()
-//keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
+keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
 android {
     namespace 'com.track.finance2up'
@@ -32,7 +32,7 @@ android {
         release {
             keyAlias keystoreProperties['KEY_ALIAS']
             keyPassword keystoreProperties['KEY_PASSWORD']
-//            storeFile file(keystoreProperties['STORE_FILE'])
+            storeFile file(keystoreProperties['STORE_FILE'])
             storePassword keystoreProperties['STORE_PASSWORD']
         }
     }
