+++ app/src/main/java/com/track/finance2up/MainActivity.kt
@@ -7,8 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
-import com.aibles.account.presentation.AccountScreen
-import com.finance2up.authentication.presentation.ui.login.LoginScreen
+import com.finance2up.authentication.presentation.ui.register.RegisterScreen
 import com.track.finance2up.ui.theme.Finance2UpTheme
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -25,7 +24,8 @@ class MainActivity : ComponentActivity() {
                 ) {
 //                    MainScreen()
 //                    LoginScreen()
-                    AccountScreen()
+//                    AccountScreen()
+                    RegisterScreen()
                 }
             }
         }
