+++ feature/account/build.gradle.kts
@@ -55,7 +55,7 @@ dependencies {
     implementation(project(Modules.common))
 
     implementation(Libraries.daggerHilt)
-    implementation(project(mapOf("path" to ":common")))
+
     kapt(Libraries.hiltCompiler)
     implementation(Libraries.hiltNavigationCompose)
