+++ feature/authentication/src/main/res/values/strings.xml
@@ -16,8 +16,8 @@
     <string name="register_title_two">An Account </string>
     <string name="register_content_description_floating_icon">next</string>
     <string name="register_Already_have_an_account">Already have an account?</string>
-    <string name="spaceeee">" "</string>
-    <string name="clear">clear</string>
+    <string name="register_space">" "</string>
+    <string name="register_clear">clear</string>
     <string name="show_password">show password</string>
     <string name="hide_password">hide password</string>
     <string name="register_name">Name</string>
@@ -33,6 +33,7 @@
     <string name="register_error_incorrect_email">Please enter valid email</string>
     <string name="register_error_incorrect_password">Please enter valid password</string>
     <string name="register_error_incorrect_confirm_password">Please enter valid confirm password</string>
+    <string name="regsiter_hide_password">hide_password</string>
 
 
 </resources>
